+++ matrix.h
@@ -5,8 +5,8 @@
 #ifndef MATRIX 
 #define MATRIX
 
-int compute_covariance_matrix(double **x,double **cov,int numberOfPoints,int numberOfDimensions);
+int compute_covariance_matrix(REAL **x,REAL **cov,int numberOfPoints,int numberOfDimensions);
 
-void find_mean(double **points,double *centroid,int numberOfPoints,int numberOfDimensions);
+void find_mean(REAL **points,REAL *centroid,int numberOfPoints,int numberOfDimensions);
 
 #endif
